perf: split trpc context from router builder
