+++ apps/web/src/pages/api/v1/[...trpc].ts
@@ -3,10 +3,10 @@ import cors from "nextjs-cors";
 import { createOpenApiNextHandler } from "trpc-to-openapi";
 
 import { appRouter } from "@kan/api";
-import { createRESTContext } from "@kan/api/trpc";
+import { createRESTContext } from "@kan/api/trpc-context";
+import { withRateLimit } from "@kan/api/utils/rateLimit";
 
 import { env } from "~/env";
-import { withRateLimit } from "@kan/api/utils/rateLimit";
 
 export default withRateLimit(
   { points: 100, duration: 60 },
